+++ deadchest-core/src/test/java/me/crylonz/deadchest/listener/LockedDropListenerTest.java
@@ -82,7 +82,7 @@ class LockedDropListenerTest {
 
     private Item lockedDrop(UUID ownerId, long expirationTime) {
         Item item = world.dropItemNaturally(new Location(world, 0, 65, 0), new ItemStack(Material.DIAMOND, 1));
-        LockedDropService.lockDrop(item, ownerId, "Steve", expirationTime);
+        LockedDropService.lockDrop(item, ownerId, "Steve", System.currentTimeMillis(), expirationTime);
         return item;
     }